code/rail-sahayak_2026_04_05-05_44/rail-sahayak_2026_04_05-05_44/gradio-data-app/requirements.txt
databricks-sdk~=0.33.0
databricks-sql-connector~=3.4.0
gradio~=4.44.0
huggingface-hub~=0.35.3
pandas~=2.0.0
pyarrow~=14.0.0
joblib~=1.3.0
scikit-learn~=1.3.0
